+++ web-common/src/features/charts/ChartsHeader.svelte
@@ -4,12 +4,12 @@
   import PanelCTA from "@rilldata/web-common/components/panel/PanelCTA.svelte";
   import GenerateVegaSpecPrompt from "@rilldata/web-common/features/charts/prompt/GenerateVegaSpecPrompt.svelte";
   import { fileArtifacts } from "@rilldata/web-common/features/entity-management/file-artifacts";
-  import { ResourceKind } from "@rilldata/web-common/features/entity-management/resource-selectors";
-  import { handleEntityRename } from "@rilldata/web-common/features/entity-management/ui-actions";
   import {
     extractFileName,
     splitFolderAndName,
-  } from "@rilldata/web-common/features/sources/extract-file-name";
+  } from "@rilldata/web-common/features/entity-management/file-path-utils";
+  import { ResourceKind } from "@rilldata/web-common/features/entity-management/resource-selectors";
+  import { handleEntityRename } from "@rilldata/web-common/features/entity-management/ui-actions";
   import { WorkspaceHeader } from "@rilldata/web-common/layout/workspace";
   import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
