+++ web-common/src/features/editor/getExtensionsForFile.ts
@@ -1,6 +1,6 @@
 import { markdown } from "@codemirror/lang-markdown";
 import { yaml } from "@rilldata/web-common/components/editor/presets/yaml";
-import { extractFileExtension } from "@rilldata/web-common/features/sources/extract-file-name";
+import { extractFileExtension } from "@rilldata/web-common/features/entity-management/file-path-utils";
 
 export const FileExtensionToEditorExtension = {
   ".yaml": yaml(),
