+++ web-common/src/features/entity-management/RenameAssetModal.svelte
@@ -4,11 +4,11 @@
   import Input from "@rilldata/web-common/components/forms/Input.svelte";
   import SubmissionError from "@rilldata/web-common/components/forms/SubmissionError.svelte";
   import { Dialog } from "@rilldata/web-common/components/modal/index";
+  import { splitFolderAndName } from "@rilldata/web-common/features/entity-management/file-path-utils";
   import {
     useAllFileNames,
     useDirectoryNamesInDirectory,
   } from "@rilldata/web-common/features/entity-management/file-selectors";
-  import { splitFolderAndName } from "@rilldata/web-common/features/sources/extract-file-name";
   import { useQueryClient } from "@tanstack/svelte-query";
   import { createForm } from "svelte-forms-lib";
   import * as yup from "yup";
