+++ web-common/src/features/entity-management/file-artifacts.ts
@@ -1,25 +1,25 @@
 import { parseKindAndNameFromFile } from "@rilldata/web-common/features/entity-management/file-content-utils";
+import { extractFileName } from "@rilldata/web-common/features/entity-management/file-path-utils";
 import {
   fetchFileContent,
   fetchMainEntityFiles,
 } from "@rilldata/web-common/features/entity-management/file-selectors";
 import {
-  fetchResources,
   ResourceKind,
+  fetchResources,
   useProjectParser,
   useResource,
 } from "@rilldata/web-common/features/entity-management/resource-selectors";
-import { extractFileName } from "@rilldata/web-common/features/sources/extract-file-name";
 import { queryClient } from "@rilldata/web-common/lib/svelte-query/globalQueryClient";
 import {
-  type V1ParseError,
   V1ReconcileStatus,
+  type V1ParseError,
   type V1Resource,
   type V1ResourceName,
 } from "@rilldata/web-common/runtime-client";
 import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
 import type { QueryClient } from "@tanstack/svelte-query";
-import { derived, get, type Readable, writable } from "svelte/store";
+import { derived, get, writable, type Readable } from "svelte/store";
 
 export class FileArtifact {
   public readonly path: string;
