+++ web-common/src/features/file-explorer/nav-entry-drag-drop-store.ts
@@ -1,7 +1,7 @@
+import { splitFolderAndName } from "@rilldata/web-common/features/entity-management/file-path-utils";
 import { ResourceKind } from "@rilldata/web-common/features/entity-management/resource-selectors";
 import { directoryState } from "@rilldata/web-common/features/file-explorer/directory-store";
 import { getPaddingFromPath } from "@rilldata/web-common/features/file-explorer/nav-tree-spacing";
-import { splitFolderAndName } from "@rilldata/web-common/features/sources/extract-file-name";
 import { get, writable } from "svelte/store";
 
 export type NavDragData = {
