+++ web-local/src/routes/(application)/[type=workspace]/[name]/+page.svelte
@@ -9,6 +9,7 @@
     FileArtifact,
     fileArtifacts,
   } from "@rilldata/web-common/features/entity-management/file-artifacts";
+  import { splitFolderAndName } from "@rilldata/web-common/features/entity-management/file-path-utils";
   import {
     ResourceKind,
     resourceIsLoading,
@@ -23,7 +24,6 @@
   import SourceEditor from "@rilldata/web-common/features/sources/editor/SourceEditor.svelte";
   import UnsavedSourceDialog from "@rilldata/web-common/features/sources/editor/UnsavedSourceDialog.svelte";
   import ErrorPane from "@rilldata/web-common/features/sources/errors/ErrorPane.svelte";
-  import { splitFolderAndName } from "@rilldata/web-common/features/sources/extract-file-name";
   import WorkspaceInspector from "@rilldata/web-common/features/sources/inspector/WorkspaceInspector.svelte";
   import {
     refreshSource,
@@ -276,7 +276,7 @@
 </script>
 
 <svelte:head>
-  <title>Rill Developer | {assetName}</title>
+  <title>Rill Developer | {fileName}</title>
 </svelte:head>
 
 {#if fileNotFound}
